refactor(profile): migrate profile store to profileSlice and update imports
