+++ src/libs/web3-data-provider/Web3Provider.tsx
@@ -11,7 +11,7 @@ import { TorusConnector } from '@web3-react/torus-connector';
 import { WalletLinkConnector } from '@web3-react/walletlink-connector';
 import { BigNumber, PopulatedTransaction, providers } from 'ethers';
 import React, { ReactElement, useCallback, useEffect, useState } from 'react';
-import { useProfileStore } from 'src/store/profile-store';
+import { useProfileStore } from 'src/store/profileSlice';
 import { useRootStore } from 'src/store/root';
 import { getNetworkConfig } from 'src/utils/marketsAndNetworksConfig';
 import { hexToAscii } from 'src/utils/utils';
